requirements.txt
streamlit
chromadb
pandas
python-dotenv
google-generativeai
pypdf
python-docx
fastapi
uvicorn
pydantic
requests